requirements.txt
psycopg2-binary
pika
pytest
pytest-html
